Fit a Cox-Type KEE Model

Fit the proportional hazards model for sparse longitudinal covariate data using a kernel estimating-equation approach.

Usage

kee_cox(formula, data, id, obs_times, h)

Arguments

formula

A model formula with a survival::Surv() response.

data

Data frame containing all variables used in the fit.

id

Subject identifier aligned row-wise with data.

obs_times

Longitudinal observation times aligned row-wise with data.

h

Positive kernel bandwidth.

Value

An object of class kee containing coefficient estimates, the estimated variance-covariance matrix, the estimating-equation matrices, convergence status, and the original function call.

Details

kee_cox() targets the proportional hazards case without estimating a nonparametric baseline component. It is therefore a useful specialized alternative to skmle() when the scientific model is Cox-type and the main interest is in the regression coefficients.

References

Sun, Dayu, Zhuowei Sun, Xingqiu Zhao, and Hongyuan Cao. "Kernel Meets Sieve: Transformed Hazards Models with Sparse Longitudinal Covariates." Journal of the American Statistical Association (2025): 1-12.

Cao, Hongyuan, et al. "Inference for Cox models with sparse longitudinal covariates." Biometrika (2015).

Examples

library(survival)

set.seed(123)
dat <- sim_skmle_data(
  n = 80,
  mu = function(tt) 8 * (0.75 + (0.5 - tt)^2),
  mu_bar = 8,
  alpha = function(tt) 0.5 * 0.75 + 0.75 * (tt * (1 - sin(2 * pi * (tt - 0.25)))),
  beta = c(1, -0.5),
  s = 0,
  cen = 0.7
)

fit_cox <- kee_cox(
  Surv(X, delta) ~ covariates,
  data = dat,
  id = id,
  obs_times = obs_times,
  h = 0.5
)

fit_cox
#> Call:
#> kee_cox(formula = Surv(X, delta) ~ covariates, data = dat, id = id, 
#>     obs_times = obs_times, h = 0.5)
#> 
#> Coefficients:
#> covariates1 covariates2 
#>   0.8514471  -0.5051830 
summary(fit_cox)
#> Call:
#> kee_cox(formula = Surv(X, delta) ~ covariates, data = dat, id = id, 
#>     obs_times = obs_times, h = 0.5)
#> 
#>   n= 80
#> 
#>             Estimate Std. Error z value Pr(>|z|)   
#> covariates1  0.85145    0.28964  2.9397 0.003286 **
#> covariates2 -0.50518    0.34852 -1.4495 0.147197   
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1